pyproject.toml
[tool.poetry]
name = "project name"
version = "0.1.0"
description = ""
authors = ["alexk1919"]

[tool.poetry.dependencies]
python = "^3.11.2"
fastapi = "^0.92.0"
motor = "^3.1.2"
gunicorn = "^20.1.0"
PyYAML = "^6.0"
uvicorn = "^0.20.0"
python-dotenv = "^1.0.0"
psutil = "^5.9.4"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.20.3"
mypy = "^1.1.1"
requests = "^2.28.2"

[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
